move BEArgs to mbe.py
