+++ src/quemb/molbe/scanner.py
@@ -8,7 +8,7 @@ import numpy as np
 from pyscf import gto, lib, scf
 
 from quemb.molbe import BE, fragmentate
-from quemb.molbe.chemfrag import ChemGenArgs
+from quemb.molbe.mbe import BEArgs
 
 
 def energy_hf(mol, fd_info=None):
@@ -144,40 +144,6 @@ def energy_be(mol, fd_info=None, be_args=None):
         )
 
     return mybe.ebe_tot
-
-
-@dataclass
-class BEArgs:
-    """Container for BE fragmentation, solver and optimization settings."""
-
-    # fragmentate keywords
-    n_BE: int = 2
-    frag_type: str = "chemgen"
-    frozen_core: bool = False
-    additional_args: ChemGenArgs | None = None
-
-    # BE initialization keywords
-    lo_method: str = "lowdin"
-    int_transform: str = "in-core"
-    auxbasis: str | None = None
-    initialize_fragment_idx: list[int] | None = None
-
-    # oneshot/optimize keywords
-    solver: str = "CCSD"
-    use_cumulant: bool = True
-    nproc: int = 1
-    ompnum: int = 4
-
-    # additional optimize keywords
-    optimize: bool = True
-    only_chem: bool = False
-    method: str = "QN"
-    conv_tol: float = 1.0e-6
-    relax_density: bool = False
-    jac_solver: str = "HF"
-    max_iter: int = 500
-    trust_region: bool = False
-    step_size: float = 1e-6
 
 
 @dataclass
